+++ modules/engine/builtin/cpp/getpidBuiltin.cpp
@@ -23,29 +23,43 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "NelsonGateway.hpp"
-#include "argvBuiltin.hpp"
-#include "getnelsonmodeBuiltin.hpp"
-#include "isquietmodeBuiltin.hpp"
 #include "getpidBuiltin.hpp"
+#include "Error.hpp"
+#include "NelsonPIDs.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
-const std::wstring gatewayName = L"engine";
-//=============================================================================
-static const nlsGateway gateway[]
-    = { { "getnelsonmode", (void*)Nelson::EngineGateway::getnelsonmodeBuiltin, 1, 0,
-            CPP_BUILTIN_WITH_EVALUATOR },
-          { "argv", (void*)Nelson::EngineGateway::argvBuiltin, 1, 0, CPP_BUILTIN_WITH_EVALUATOR },
-          { "isquietmode", (void*)Nelson::EngineGateway::isquietmodeBuiltin, 1, 0,
-              CPP_BUILTIN_WITH_EVALUATOR },
-          { "getpid", (void*)Nelson::EngineGateway::getpidBuiltin, 1, 0, CPP_BUILTIN } };
-//=============================================================================
-NLSGATEWAYFUNC(gateway)
-//=============================================================================
-NLSGATEWAYINFO(gateway)
-//=============================================================================
-NLSGATEWAYREMOVE(gateway)
-//=============================================================================
-NLSGATEWAYNAME()
+ArrayOfVector
+Nelson::EngineGateway::getpidBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    if (nLhs > 1) {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    }
+    switch (argIn.size()) {
+    case 0: {
+        retval.push_back(ArrayOf::doubleConstructor((double)getCurrentPID()));
+    } break;
+    case 1: {
+        std::wstring param = argIn[0].getContentAsWideString();
+        if (param == L"running") {
+            std::vector<int> pids = getNelsonPIDs();
+            ArrayOf res;
+            Dimensions dims(1, pids.size());
+            double* pd = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, pids.size());
+            res = ArrayOf(NLS_DOUBLE, dims, pd);
+            for (indexType k = 0; k < dims.getElementCount(); ++k) {
+                pd[k] = (double)pids[k];
+            }
+            retval.push_back(res);
+        } else {
+            Error(_("Wrong value for #1 argument: 'running' expected."));
+        }
+    } break;
+    default: {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    } break;
+    }
+    return retval;
+}
 //=============================================================================